+++ src/ria/hid/kbd.h
@@ -15,13 +15,13 @@ void kbd_task(void);
 void kbd_stop(void);
 
 // Parse HID report descriptor
-bool kbd_mount(uint8_t slot, uint8_t const *desc_data, uint16_t desc_len);
+bool kbd_mount(int slot, uint8_t const *desc_data, uint16_t desc_len);
 
 // Clean up descriptor when device is disconnected.
-bool kbd_umount(uint8_t slot);
+bool kbd_umount(int slot);
 
 // Process HID keyboard report.
-void kbd_report(uint8_t slot, uint8_t const *data, size_t size);
+void kbd_report(int slot, uint8_t const *data, size_t size);
 
 // Set the extended register value.
 bool kbd_xreg(uint16_t word);
